% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeClipboard.r
\name{cbWrite}
\alias{cbWrite}
\title{A function for copying text to the clipboard on macOS. Note that this is very simple...it doesn't allow for different separators, etc. That could be added fairly easily, but I haven't found the need yet.}
\usage{
cbWrite(x)
}
\arguments{
\item{x}{is the data to be written to the clipboard. I have not extensively tested what can and can't be written at this point. Since the function calls write.table, the help files for it are likely a good start.}
}
\description{
A function for copying text to the clipboard on macOS. Note that this is very simple...it doesn't allow for different separators, etc. That could be added fairly easily, but I haven't found the need yet.
}
