% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_low.R
\name{ci_low}
\alias{ci_low}
\title{Calculate the lower bound of the 95% confidence interval given a standard error.}
\usage{
ci_low(mean_var, se_var)
}
\arguments{
\item{mean_var}{is mean of the data used to calculate the CI}

\item{se_var}{is the se (typically calculated with jscTools::stderr) used to calculate the CI}
}
\description{
Calculate the lower bound of the 95% confidence interval given a standard error.
}
\examples{
ci_low(mean_var, se_var)
}
