% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decruftQ.R
\name{decruftQ}
\alias{decruftQ}
\title{A function for finding the standard error}
\usage{
decruftQ(input_csv, skip = 3)
}
\arguments{
\item{input_csv}{is the downloaded Qualtrics csv with 3 lines of cruft at the top}
}
\description{
This function takes a series of numbers (vector, etc.) and finds the standard error. See http://www.jscarlton.net/post/2017-04-04QualtricsImport/ for more information.
}
\examples{
decruftQ()
}
