\name{jscTools-package}
\alias{jscTools-package}
\alias{jscTools}
\docType{package}
\title{
\packageTitle{jscTools}
}
\description{
\packageDescription{jscTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jscTools}
\packageIndices{jscTools}
A random assortment of functions and other tools for working with R.
}
\author{
\packageAuthor{jscTools}

Maintainer: \packageMaintainer{jscTools}
}

\keyword{ package }

\examples{
~~ simple examples of the most important functions ~~
}
