% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderr.R
\name{stderr}
\alias{stderr}
\title{A function for finding the standard error}
\usage{
stderr(x)
}
\arguments{
\item{x}{is a list of numbers}
}
\description{
This function takes a series of numbers (vector, etc.) and finds the standard error.
}
\examples{
stderr()
}
