% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{theme_fira}
\alias{theme_fira}
\title{theme_fira}
\usage{
theme_fira(
  base_family = "Fira Mono",
  base_size = 12,
  title_family = "Fira Sans",
  plot_title_family = title_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = title_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = title_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = title_family,
  caption_size = 6,
  caption_face = "plain",
  caption_margin = 10,
  axis_title_family = title_family,
  axis_title_size = 12,
  axis_title_face = "plain",
  axis_title_just = "ct",
  legend_title_family = title_family,
  legend_title_size = 10,
  legend_title_face = "plain",
  plot_margin = ggplot2::margin(1, 1, 1, 1),
  panel_spacing = ggplot2::unit(0.25, "lines"),
  grid = TRUE,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\description{
I will eventually create my own theme out of this, but for now it's a thinly disguised version of Kieran Healy's Myriad theme.
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_fira()

# seminal bar chart

update_geom_font_defaults(family=font_fira_light)

count(mpg, class) \%>\%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label=n), nudge_y=3) +
  labs(x="Fuel effiiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_fira(grid="Y") +
  theme(axis.text.y=element_blank())
}
}
