% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hmr.R
\name{theme_hmr}
\alias{theme_hmr}
\title{A theme for plots for the Half Moon Reef paper
Based on ggthemes' fivethirtyeight, but only in that I used that as a shell}
\usage{
theme_hmr(
  base_size = 12,
  base_family = "Whitney-Medium",
  title_family = "Whitney-Semibold"
)
}
\arguments{
\item{base_size}{Base font size for the plot. Defaults to 12.}

\item{base_family}{Base font family. Defaults to Whitney Medium.}

\item{title_family}{Family for the title font. Defaults to Whitney Semibold}
}
\description{
A theme for plots for the Half Moon Reef paper
Based on ggthemes' fivethirtyeight, but only in that I used that as a shell
}
