% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_cloud.R
\name{transcript_cloud}
\alias{transcript_cloud}
\title{A function for creating a wordcloud from a text transcript}
\usage{
transcript_cloud(
  input_path,
  save_path,
  cloud_width = 2000,
  cloud_height = 2000,
  cloud_res = 300,
  cloud_max_words = 100,
  color_pal = c(brewer.pal(5, "Blues")[3], brewer.pal(5, "Blues")[4], brewer.pal(5,
    "Blues")[5])
)
}
\arguments{
\item{input_path}{The path of the text file to turn into a wordcloud}

\item{save_path}{The path where you want the saved image to go. Should end in ".png".}

\item{cloud_width}{The width of your saved wordcloud image, in pixels. Defaults to 2000.}

\item{cloud_height}{The height of your saved wordcloud image, in pixels. Defaults to 2000.}

\item{cloud_res}{The resolution of your saved wordcloud image, in dpi Defaults to 300}

\item{cloud_max_words}{The maximum number of words you want to plot in your wordcloud. Defaults to 100.}

\item{color_pal}{The color palette that you want to use for your wordcloud. Choose one color if you want a monochrome wordcloud or a series of colors if you want it to vary the color by the relative frequency of the word. See the wordcloud documentation for details. Defaults to c(brewer.pal(5,"Blues")[3],brewer.pal(5,"Blues")[4],brewer.pal(5,"Blues")[5])}
}
\description{
A function for creating a wordcloud from a text transcript
}
\keyword{analysis,}
\keyword{text}
\keyword{wordcloud}
