% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCloudExport.r
\name{wordCloudExport}
\alias{wordCloudExport}
\title{A function for exporting a word cloud, using the wordcloud package. Requires the following libraries: wordcloud, tm, SnowballC}
\usage{
wordCloudExport(
  cloudCorpus,
  exportName = "wordcloud.png",
  cloudStopWords = stopwords("english"),
  cloudRandomOrder = FALSE,
  cloudRotPer = 0,
  cloudMaxWords = Inf,
  fileWidth = 2000,
  fileHeight = 2000,
  fileRes = 300
)
}
\arguments{
\item{cloudCorpus}{the corpus to be turned into a word cloud. I created it expecting a column of responses to an open-ended survey question.}

\item{exportName}{the name of the file you want to export. Defaults to wordcloud.png}

\item{cloudStopWords}{a list of stopwords for the word cloud. Defaults to stopwords("english"). Can add more stopwords with newStopWords <- c("foo", stopwords("english")).}
}
\description{
This function takes a series of numbers (vector, etc.) and finds the standard error.
}
\examples{
wordCloudExport()

}
